+++ packages/ckeditor5-undo/tests/undoediting-integration.js
@@ -1,5 +1,5 @@
 /**
- * @license Copyright (c) 2003-2017, CKSource - Frederico Knabben. All rights reserved.
+ * @license Copyright (c) 2003-2018, CKSource - Frederico Knabben. All rights reserved.
  * For licensing, see LICENSE.md.
  */
 
@@ -9,53 +9,57 @@ import ClassicEditor from '@ckeditor/ckeditor5-editor-classic/src/classiceditor'
 
 import Range from '@ckeditor/ckeditor5-engine/src/model/range';
 import Position from '@ckeditor/ckeditor5-engine/src/model/position';
-import Element from '@ckeditor/ckeditor5-engine/src/model/element';
-import UndoEngine from '../src/undoengine';
+import Batch from '@ckeditor/ckeditor5-engine/src/model/batch';
+import UndoEditing from '../src/undoediting';
 
 import Paragraph from '@ckeditor/ckeditor5-paragraph/src/paragraph';
-import HeadingEngine from '@ckeditor/ckeditor5-heading/src/headingengine';
+import HeadingEditing from '@ckeditor/ckeditor5-heading/src/headingediting';
 import Typing from '@ckeditor/ckeditor5-typing/src/typing';
 import Enter from '@ckeditor/ckeditor5-enter/src/enter';
 import Clipboard from '@ckeditor/ckeditor5-clipboard/src/clipboard';
-import BoldEngine from '@ckeditor/ckeditor5-basic-styles/src/boldengine';
+import BoldEditing from '@ckeditor/ckeditor5-basic-styles/src/bold/boldediting';
 
-import buildModelConverter from '@ckeditor/ckeditor5-engine/src/conversion/buildmodelconverter';
-import buildViewConverter from '@ckeditor/ckeditor5-engine/src/conversion/buildviewconverter';
+import { downcastElementToElement } from '@ckeditor/ckeditor5-engine/src/conversion/downcast-converters';
+import { upcastElementToElement } from '@ckeditor/ckeditor5-engine/src/conversion/upcast-converters';
 
 import { setData, getData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model';
 
-describe( 'UndoEngine integration', () => {
-	let editor, doc, root, div;
+describe( 'UndoEditing integration', () => {
+	let editor, model, doc, root, div;
 
 	beforeEach( () => {
 		div = document.createElement( 'div' );
 		document.body.appendChild( div );
 
-		return ClassicEditor.create( div, { plugins: [ Paragraph, HeadingEngine, Typing, Enter, Clipboard, BoldEngine, UndoEngine ] } )
+		return ClassicEditor.create( div, { plugins: [ Paragraph, HeadingEditing, Typing, Enter, Clipboard, BoldEditing, UndoEditing ] } )
 			.then( newEditor => {
 				editor = newEditor;
 
-				doc = editor.document;
+				model = editor.model;
+				doc = model.document;
 
 				// Add "div feature".
-				doc.schema.registerItem( 'div', '$block' );
-				buildModelConverter().for( editor.data.modelToView, editor.editing.modelToView ).fromElement( 'div' ).toElement( 'div' );
-				buildViewConverter().for( editor.data.viewToModel ).fromElement( 'div' ).toElement( 'div' );
+				model.schema.register( 'div', { inheritAllFrom: '$block' } );
+
+				editor.conversion.for( 'downcast' ).add( downcastElementToElement( { model: 'div', view: 'div' } ) );
+				editor.conversion.for( 'upcast' ).add( upcastElementToElement( { model: 'div', view: 'div' } ) );
 
 				root = doc.getRoot();
 			} );
 	} );
 
 	function setSelection( pathA, pathB ) {
-		doc.selection.setRanges( [ new Range( new Position( root, pathA ), new Position( root, pathB ) ) ] );
+		model.change( writer => {
+			writer.setSelection( new Range( new Position( root, pathA ), new Position( root, pathB ) ) );
+		} );
 	}
 
 	function input( input ) {
-		setData( doc, input );
+		setData( model, input );
 	}
 
 	function output( output ) {
-		expect( getData( doc ) ).to.equal( output );
+		expect( getData( model ) ).to.equal( output );
 	}
 
 	function undoDisabled() {
@@ -70,8 +74,8 @@ describe( 'UndoEngine integration', () => {
 		it( 'add and undo', () => {
 			input( '<paragraph>fo[]o</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().insert( doc.selection.getFirstPosition(), 'zzz' );
+			model.change( writer => {
+				writer.insertText( 'zzz', doc.selection.getFirstPosition() );
 			} );
 			output( '<paragraph>fozzz[]o</paragraph><paragraph>bar</paragraph>' );
 
@@ -84,17 +88,16 @@ describe( 'UndoEngine integration', () => {
 		it( 'multiple adding and undo', () => {
 			input( '<paragraph>fo[]o</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch()
-					.insert( doc.selection.getFirstPosition(), 'zzz' )
-					.insert( new Position( root, [ 1, 0 ] ), 'xxx' );
+			model.change( writer => {
+				writer.insertText( 'zzz', doc.selection.getFirstPosition() );
+				writer.insertText( 'xxx', new Position( root, [ 1, 0 ] ) );
 			} );
 
 			output( '<paragraph>fozzz[]o</paragraph><paragraph>xxxbar</paragraph>' );
 
-			doc.enqueueChanges( () => {
+			model.change( writer => {
 				setSelection( [ 1, 0 ], [ 1, 0 ] );
-				doc.batch().insert( doc.selection.getFirstPosition(), 'yyy' );
+				writer.insertText( 'yyy', doc.selection.getFirstPosition() );
 			} );
 
 			output( '<paragraph>fozzzo</paragraph><paragraph>yyy[]xxxbar</paragraph>' );
@@ -111,14 +114,14 @@ describe( 'UndoEngine integration', () => {
 		it( 'multiple adding mixed with undo', () => {
 			input( '<paragraph>fo[]o</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().insert( doc.selection.getFirstPosition(), 'zzz' );
+			model.change( writer => {
+				writer.insertText( 'zzz', doc.selection.getFirstPosition() );
 			} );
 			output( '<paragraph>fozzz[]o</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
+			model.change( writer => {
 				setSelection( [ 1, 0 ], [ 1, 0 ] );
-				doc.batch().insert( doc.selection.getFirstPosition(), 'yyy' );
+				writer.insertText( 'yyy', doc.selection.getFirstPosition() );
 			} );
 
 			output( '<paragraph>fozzzo</paragraph><paragraph>yyy[]bar</paragraph>' );
@@ -126,9 +129,9 @@ describe( 'UndoEngine integration', () => {
 			editor.execute( 'undo' );
 			output( '<paragraph>fozzzo</paragraph><paragraph>[]bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
+			model.change( writer => {
 				setSelection( [ 0, 0 ], [ 0, 0 ] );
-				doc.batch().insert( doc.selection.getFirstPosition(), 'xxx' );
+				writer.insertText( 'xxx', doc.selection.getFirstPosition() );
 			} );
 			output( '<paragraph>xxx[]fozzzo</paragraph><paragraph>bar</paragraph>' );
 
@@ -144,14 +147,14 @@ describe( 'UndoEngine integration', () => {
 		it( 'multiple remove and undo', () => {
 			input( '<paragraph>[]foo</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().remove( Range.createFromPositionAndShift( doc.selection.getFirstPosition(), 2 ) );
+			model.change( writer => {
+				writer.remove( Range.createFromPositionAndShift( doc.selection.getFirstPosition(), 2 ) );
 			} );
 			output( '<paragraph>[]o</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
+			model.change( writer => {
 				setSelection( [ 1, 1 ], [ 1, 1 ] );
-				doc.batch().remove( Range.createFromPositionAndShift( doc.selection.getFirstPosition(), 2 ) );
+				writer.remove( Range.createFromPositionAndShift( doc.selection.getFirstPosition(), 2 ) );
 			} );
 			output( '<paragraph>o</paragraph><paragraph>b[]</paragraph>' );
 
@@ -169,14 +172,14 @@ describe( 'UndoEngine integration', () => {
 		it( 'add and remove different parts and undo', () => {
 			input( '<paragraph>fo[]o</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().insert( doc.selection.getFirstPosition(), 'zzz' );
+			model.change( writer => {
+				writer.insertText( 'zzz', doc.selection.getFirstPosition() );
 			} );
 			output( '<paragraph>fozzz[]o</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
+			model.change( writer => {
 				setSelection( [ 1, 2 ], [ 1, 2 ] );
-				doc.batch().remove( Range.createFromPositionAndShift( new Position( root, [ 1, 1 ] ), 1 ) );
+				writer.remove( Range.createFromPositionAndShift( new Position( root, [ 1, 1 ] ), 1 ) );
 			} );
 			output( '<paragraph>fozzzo</paragraph><paragraph>b[]r</paragraph>' );
 
@@ -192,13 +195,13 @@ describe( 'UndoEngine integration', () => {
 		it( 'add and remove same part and undo', () => {
 			input( '<paragraph>fo[]o</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().insert( doc.selection.getFirstPosition(), 'zzz' );
+			model.change( writer => {
+				writer.insertText( 'zzz', doc.selection.getFirstPosition() );
 			} );
 			output( '<paragraph>fozzz[]o</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().remove( Range.createFromPositionAndShift( new Position( root, [ 0, 2 ] ), 3 ) );
+			model.change( writer => {
+				writer.remove( Range.createFromPositionAndShift( new Position( root, [ 0, 2 ] ), 3 ) );
 			} );
 			output( '<paragraph>fo[]o</paragraph><paragraph>bar</paragraph>' );
 
@@ -214,8 +217,8 @@ describe( 'UndoEngine integration', () => {
 		it( 'undo remove all content', () => {
 			input( '<paragraph>foo[]</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().remove( Range.createIn( root ) );
+			model.change( writer => {
+				writer.remove( Range.createIn( root ) );
 			} );
 			output( '<paragraph>[]</paragraph>' ); // All hail our king and savior, autoparagraphing!
 
@@ -227,14 +230,17 @@ describe( 'UndoEngine integration', () => {
 
 		it( 'undo insert first content', () => {
 			input( '' );
+			output( '<paragraph>[]</paragraph>' ); // All hail our king and savior, autoparagraphing!
 
-			doc.enqueueChanges( () => {
-				doc.batch().insert( doc.selection.getFirstPosition(), new Element( 'heading1' ) );
+			model.change( writer => {
+				writer.remove( Range.createIn( root ) );
+				writer.insertElement( 'heading1', doc.selection.getFirstPosition() );
 			} );
+
 			output( '<heading1>[]</heading1>' );
 
 			editor.execute( 'undo' );
-			output( '<paragraph>[]</paragraph>' ); // All hail our king and savior, autoparagraphing!
+			output( '<paragraph>[]</paragraph>' );
 
 			undoDisabled();
 		} );
@@ -246,11 +252,10 @@ describe( 'UndoEngine integration', () => {
 			const p = root.getChild( 0 );
 			const pos = new Position( root, [ 0 ] );
 
-			doc.enqueueChanges( () => {
-				doc.batch()
-					.remove( p )
-					.insert( pos, new Element( 'heading1' ) )
-					.insert( pos.getShiftedBy( 1 ), new Element( 'heading2' ) );
+			model.change( writer => {
+				writer.remove( p );
+				writer.insertElement( 'heading1', pos );
+				writer.insertElement( 'heading2', pos.getShiftedBy( 1 ) );
 			} );
 
 			output( '<heading1>[]</heading1><heading2></heading2>' );
@@ -267,13 +272,13 @@ describe( 'UndoEngine integration', () => {
 		it( 'move same content twice then undo', () => {
 			input( '<paragraph>f[o]z</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().move( doc.selection.getFirstRange(), new Position( root, [ 1, 0 ] ) );
+			model.change( writer => {
+				writer.move( doc.selection.getFirstRange(), new Position( root, [ 1, 0 ] ) );
 			} );
 			output( '<paragraph>fz</paragraph><paragraph>[o]bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().move( doc.selection.getFirstRange(), new Position( root, [ 0, 2 ] ) );
+			model.change( writer => {
+				writer.move( doc.selection.getFirstRange(), new Position( root, [ 0, 2 ] ) );
 			} );
 			output( '<paragraph>fz[o]</paragraph><paragraph>bar</paragraph>' );
 
@@ -289,14 +294,14 @@ describe( 'UndoEngine integration', () => {
 		it( 'move content and new parent then undo', () => {
 			input( '<paragraph>f[o]z</paragraph><paragraph>bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().move( doc.selection.getFirstRange(), new Position( root, [ 1, 0 ] ) );
+			model.change( writer => {
+				writer.move( doc.selection.getFirstRange(), new Position( root, [ 1, 0 ] ) );
 			} );
 			output( '<paragraph>fz</paragraph><paragraph>[o]bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
+			model.change( writer => {
 				setSelection( [ 1 ], [ 2 ] );
-				doc.batch().move( doc.selection.getFirstRange(), new Position( root, [ 0 ] ) );
+				writer.move( doc.selection.getFirstRange(), new Position( root, [ 0 ] ) );
 			} );
 			output( '[<paragraph>obar</paragraph>]<paragraph>fz</paragraph>' );
 
@@ -314,24 +319,26 @@ describe( 'UndoEngine integration', () => {
 		it( 'attributes then insert inside then undo', () => {
 			input( '<paragraph>fo[ob]ar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().setAttribute( doc.selection.getFirstRange(), 'bold', true );
+			model.change( writer => {
+				writer.setAttribute( 'bold', true, doc.selection.getFirstRange() );
 			} );
 			output( '<paragraph>fo[<$text bold="true">ob</$text>]ar</paragraph>' );
+			expect( doc.selection.getAttribute( 'bold' ) ).to.be.true;
 
-			doc.enqueueChanges( () => {
+			model.change( writer => {
 				setSelection( [ 0, 3 ], [ 0, 3 ] );
-				doc.batch().insert( doc.selection.getFirstPosition(), 'zzz' );
+				writer.insertText( 'zzz', doc.selection.getFirstPosition() );
 			} );
-			output( '<paragraph>fo<$text bold="true">o</$text>zzz<$text bold="true">[]b</$text>ar</paragraph>' );
-			expect( doc.selection.getAttribute( 'bold' ) ).to.true;
+			output( '<paragraph>fo<$text bold="true">o</$text>zzz[]<$text bold="true">b</$text>ar</paragraph>' );
+			expect( doc.selection.getAttribute( 'bold' ) ).to.be.undefined;
 
 			editor.execute( 'undo' );
 			output( '<paragraph>fo<$text bold="true">o[]b</$text>ar</paragraph>' );
-			expect( doc.selection.getAttribute( 'bold' ) ).to.true;
+			expect( doc.selection.getAttribute( 'bold' ) ).to.be.true;
 
 			editor.execute( 'undo' );
 			output( '<paragraph>fo[ob]ar</paragraph>' );
+			expect( doc.selection.getAttribute( 'bold' ) ).to.be.undefined;
 
 			undoDisabled();
 		} );
@@ -339,11 +346,11 @@ describe( 'UndoEngine integration', () => {
 
 	describe( 'wrapping, unwrapping, merging, splitting', () => {
 		it( 'wrap and undo', () => {
-			doc.schema.allow( { name: '$text', inside: '$root' } );
+			model.schema.extend( '$text', { allowIn: '$root' } );
 			input( 'fo[zb]ar' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().wrap( doc.selection.getFirstRange(), 'paragraph' );
+			model.change( writer => {
+				writer.wrap( doc.selection.getFirstRange(), 'paragraph' );
 			} );
 			output( 'fo<paragraph>[zb]</paragraph>ar' );
 
@@ -354,18 +361,18 @@ describe( 'UndoEngine integration', () => {
 		} );
 
 		it( 'wrap, move and undo', () => {
-			doc.schema.allow( { name: '$text', inside: '$root' } );
+			model.schema.extend( '$text', { allowIn: '$root' } );
 			input( 'fo[zb]ar' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().wrap( doc.selection.getFirstRange(), 'paragraph' );
+			model.change( writer => {
+				writer.wrap( doc.selection.getFirstRange(), 'paragraph' );
 			} );
 			// Would be better if selection was inside P.
 			output( 'fo<paragraph>[zb]</paragraph>ar' );
 
-			doc.enqueueChanges( () => {
+			model.change( writer => {
 				setSelection( [ 2, 0 ], [ 2, 1 ] );
-				doc.batch().move( doc.selection.getFirstRange(), new Position( root, [ 0 ] ) );
+				writer.move( doc.selection.getFirstRange(), new Position( root, [ 0 ] ) );
 			} );
 			output( '[z]fo<paragraph>b</paragraph>ar' );
 
@@ -381,8 +388,8 @@ describe( 'UndoEngine integration', () => {
 		it( 'unwrap and undo', () => {
 			input( '<paragraph>foo[]bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().unwrap( doc.selection.getFirstPosition().parent );
+			model.change( writer => {
+				writer.unwrap( doc.selection.getFirstPosition().parent );
 			} );
 			output( 'foo[]bar' );
 
@@ -395,8 +402,8 @@ describe( 'UndoEngine integration', () => {
 		it( 'merge and undo', () => {
 			input( '<paragraph>foo</paragraph><paragraph>[]bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().merge( new Position( root, [ 1 ] ) );
+			model.change( writer => {
+				writer.merge( new Position( root, [ 1 ] ) );
 				// Because selection is stuck with <paragraph> it ends up in graveyard. We have to manually move it to correct node.
 				setSelection( [ 0, 3 ], [ 0, 3 ] );
 			} );
@@ -411,8 +418,8 @@ describe( 'UndoEngine integration', () => {
 		it( 'split and undo', () => {
 			input( '<paragraph>foo[]bar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				doc.batch().split( doc.selection.getFirstPosition() );
+			model.change( writer => {
+				writer.split( doc.selection.getFirstPosition() );
 				// Because selection is stuck with <paragraph> it ends up in wrong node. We have to manually move it to correct node.
 				setSelection( [ 1, 0 ], [ 1, 0 ] );
 			} );
@@ -770,13 +777,19 @@ describe( 'UndoEngine integration', () => {
 		it( 'rename leaks to other elements on undo #1', () => {
 			input( '<heading1>[]Foo</heading1><paragraph>Bar</paragraph>' );
 
-			doc.batch().rename( root.getChild( 0 ), 'paragraph' );
+			model.change( writer => {
+				writer.rename( root.getChild( 0 ), 'paragraph' );
+			} );
 			output( '<paragraph>[]Foo</paragraph><paragraph>Bar</paragraph>' );
 
-			doc.batch().split( Position.createAt( root.getChild( 0 ), 1 ) );
+			model.change( writer => {
+				writer.split( Position.createAt( root.getChild( 0 ), 1 ) );
+			} );
 			output( '<paragraph>[]F</paragraph><paragraph>oo</paragraph><paragraph>Bar</paragraph>' );
 
-			doc.batch().merge( Position.createAt( root, 2 ) );
+			model.change( writer => {
+				writer.merge( Position.createAt( root, 2 ) );
+			} );
 			output( '<paragraph>[]F</paragraph><paragraph>ooBar</paragraph>' );
 
 			editor.execute( 'undo' );
@@ -793,10 +806,14 @@ describe( 'UndoEngine integration', () => {
 		it( 'rename leaks to other elements on undo #2', () => {
 			input( '<heading1>[]Foo</heading1><paragraph>Bar</paragraph>' );
 
-			doc.batch().rename( root.getChild( 0 ), 'heading2' );
+			model.change( writer => {
+				writer.rename( root.getChild( 0 ), 'heading2' );
+			} );
 			output( '<heading2>[]Foo</heading2><paragraph>Bar</paragraph>' );
 
-			doc.batch().merge( Position.createAt( root, 1 ) );
+			model.change( writer => {
+				writer.merge( Position.createAt( root, 1 ) );
+			} );
 			output( '<heading2>[]FooBar</heading2>' );
 
 			editor.execute( 'undo' );
@@ -810,10 +827,14 @@ describe( 'UndoEngine integration', () => {
 		it( 'merge, rename, undo, undo is correct', () => {
 			input( '<heading1>[]Foo</heading1><paragraph>Bar</paragraph>' );
 
-			doc.batch().merge( Position.createAt( root, 1 ) );
+			model.change( writer => {
+				writer.merge( Position.createAt( root, 1 ) );
+			} );
 			output( '<heading1>[]FooBar</heading1>' );
 
-			doc.batch().rename( root.getChild( 0 ), 'heading2' );
+			model.change( writer => {
+				writer.rename( root.getChild( 0 ), 'heading2' );
+			} );
 			output( '<heading2>[]FooBar</heading2>' );
 
 			editor.execute( 'undo' );
@@ -827,10 +848,14 @@ describe( 'UndoEngine integration', () => {
 		it( 'wrap, split, undo, undo is correct', () => {
 			input( '<paragraph>[]Foo</paragraph><paragraph>Bar</paragraph>' );
 
-			doc.batch().wrap( Range.createIn( root ), 'div' );
+			model.change( writer => {
+				writer.wrap( Range.createIn( root ), 'div' );
+			} );
 			output( '<div><paragraph>[]Foo</paragraph><paragraph>Bar</paragraph></div>' );
 
-			doc.batch().split( new Position( root, [ 0, 0, 1 ] ) );
+			model.change( writer => {
+				writer.split( new Position( root, [ 0, 0, 1 ] ) );
+			} );
 			output( '<div><paragraph>[]F</paragraph><paragraph>oo</paragraph><paragraph>Bar</paragraph></div>' );
 
 			editor.execute( 'undo' );
@@ -877,10 +902,10 @@ describe( 'UndoEngine integration', () => {
 
 			editor.execute( 'enter' );
 
-			doc.enqueueChanges( () => {
+			model.change( writer => {
 				const range = new Range( new Position( root, [ 0, 3 ] ), new Position( root, [ 1, 3 ] ) );
 
-				doc.selection.setRanges( [ range ] );
+				writer.setSelection( range );
 
 				editor.execute( 'delete' );
 			} );
@@ -895,9 +920,9 @@ describe( 'UndoEngine integration', () => {
 		} );
 	} );
 
-	describe( 'pasting', () => {
+	describe( 'clipboard', () => {
 		function pasteHtml( editor, html ) {
-			editor.editing.view.fire( 'paste', {
+			editor.editing.view.document.fire( 'paste', {
 				dataTransfer: createDataTransfer( { 'text/html': html } ),
 				preventDefault() {}
 			} );
@@ -907,23 +932,24 @@ describe( 'UndoEngine integration', () => {
 			return {
 				getData( type ) {
 					return data[ type ];
-				}
+				},
+				setData() {}
 			};
 		}
 
 		// ckeditor5-engine#t/1065
 		it( 'undo paste into non empty element should not throw and be correct', () => {
-			doc.enqueueChanges( () => {
+			model.change( () => {
 				input( '<paragraph>Foo[]</paragraph>' );
 			} );
 
-			doc.enqueueChanges( () => {
+			model.change( () => {
 				pasteHtml( editor, '<p>a</p><p>b</p>' );
 			} );
 
 			output( '<paragraph>Fooa</paragraph><paragraph>b[]</paragraph>' );
 
-			doc.enqueueChanges( () => {
+			model.change( () => {
 				pasteHtml( editor, '<p>c</p><p>d</p>' );
 			} );
 
@@ -935,15 +961,29 @@ describe( 'UndoEngine integration', () => {
 			editor.execute( 'undo' );
 			output( '<paragraph>Foo[]</paragraph>' );
 		} );
+
+		// ckeditor5#781
+		it( 'cutting should not create empty undo step', () => {
+			input( '<paragraph>Fo[oba]r</paragraph>' );
+
+			editor.editing.view.document.fire( 'cut', {
+				dataTransfer: createDataTransfer(),
+				preventDefault() {},
+				method: 'cut'
+			} );
+
+			editor.execute( 'undo' );
+
+			output( '<paragraph>Fo[oba]r</paragraph>' );
+			undoDisabled();
+		} );
 	} );
 
 	describe( 'other edge cases', () => {
 		it( 'deleteContent between two nodes', () => {
 			input( '<paragraph>fo[o</paragraph><paragraph>b]ar</paragraph>' );
 
-			doc.enqueueChanges( () => {
-				editor.data.deleteContent( doc.selection, doc.batch() );
-			} );
+			editor.model.deleteContent( doc.selection );
 			output( '<paragraph>fo[]ar</paragraph>' );
 
 			editor.execute( 'undo' );
@@ -957,9 +997,12 @@ describe( 'UndoEngine integration', () => {
 			const gy = doc.graveyard;
 			const p = doc.getRoot().getChild( 0 );
 
-			doc.enqueueChanges( () => {
-				doc.batch().remove( p );
-				doc.batch().setAttribute( p, 'bold', true );
+			model.change( writer => {
+				writer.remove( p );
+			} );
+
+			model.change( writer => {
+				writer.setAttribute( 'bold', true, p );
 			} );
 
 			editor.execute( 'undo' );
@@ -976,20 +1019,22 @@ describe( 'UndoEngine integration', () => {
 			input( '<paragraph>Foo</paragraph><paragraph>Bar</paragraph>' );
 
 			// Remove children from graveyard because they are inserted there after `input` call.
-			doc.graveyard.removeChildren( 0, doc.graveyard.childCount );
-
-			const batchWithMerge = doc.batch();
-
-			doc.enqueueChanges( () => {
-				batchWithMerge.merge( new Position( root, [ 1 ] ) );
-
-				const split = batchWithMerge.deltas[ 0 ].getReversed();
-				const batch = doc.batch();
-				batch.addDelta( split );
-
-				doc.applyOperation( split.operations[ 0 ] );
-				doc.applyOperation( split.operations[ 1 ] );
+			model.change( writer => {
+				writer.remove( Range.createIn( doc.graveyard ) );
 			} );
+
+			const batchWithMerge = new Batch();
+
+			model.enqueueChange( batchWithMerge, writer => {
+				writer.merge( new Position( root, [ 1 ] ) );
+			} );
+
+			const split = batchWithMerge.deltas[ 0 ].getReversed();
+			const batch = new Batch();
+			batch.addDelta( split );
+
+			model.applyOperation( split.operations[ 0 ] );
+			model.applyOperation( split.operations[ 1 ] );
 
 			output( '<paragraph>[]Foo</paragraph><paragraph>Bar</paragraph>' );